Improve the OpenAPI normalizer
