+++ src/Bridge/Symfony/Bundle/Resources/config/swagger.xml
@@ -39,34 +39,6 @@
             <tag name="serializer.normalizer" priority="17" />
         </service>
 
-        <service id="api_platform.openapi.normalizer.documentation" class="ApiPlatform\Core\OpenApi\Serializer\DocumentationNormalizer" public="false">
-            <argument type="service" id="api_platform.metadata.resource.metadata_factory" />
-            <argument type="service" id="api_platform.metadata.property.name_collection_factory" />
-            <argument type="service" id="api_platform.metadata.property.metadata_factory" />
-            <argument type="service" id="api_platform.resource_class_resolver" />
-            <argument type="service" id="api_platform.operation_method_resolver" />
-            <argument type="service" id="api_platform.operation_path_resolver" />
-            <argument type="service" id="api_platform.filter_locator" />
-            <argument type="service" id="api_platform.name_converter" on-invalid="ignore" />
-            <argument>%api_platform.oauth.enabled%</argument>
-            <argument>%api_platform.oauth.type%</argument>
-            <argument>%api_platform.oauth.flow%</argument>
-            <argument>%api_platform.oauth.tokenUrl%</argument>
-            <argument>%api_platform.oauth.authorizationUrl%</argument>
-            <argument>%api_platform.oauth.scopes%</argument>
-            <argument>%api_platform.swagger.api_keys%</argument>
-            <argument type="service" id="api_platform.subresource_operation_factory" />
-            <argument>%api_platform.collection.pagination.enabled%</argument>
-            <argument>%api_platform.collection.pagination.page_parameter_name%</argument>
-            <argument>%api_platform.collection.pagination.client_items_per_page%</argument>
-            <argument>%api_platform.collection.pagination.items_per_page_parameter_name%</argument>
-            <argument type="service" id="api_platform.formats_provider" />
-            <argument>%api_platform.collection.pagination.client_enabled%</argument>
-            <argument>%api_platform.collection.pagination.enabled_parameter_name%</argument>
-            <argument type="service" id="api_platform.swagger.normalizer.api_gateway" />
-            <tag name="serializer.normalizer" priority="18" />
-        </service>
-
         <service id="api_platform.swagger.command.swagger_command" class="ApiPlatform\Core\Bridge\Symfony\Bundle\Command\SwaggerCommand">
             <argument type="service" id="api_platform.serializer" />
             <argument type="service" id="api_platform.metadata.resource.name_collection_factory" />
