+++ tests/Swagger/Serializer/DocumentationNormalizerV3Test.php
@@ -18,7 +18,6 @@ use ApiPlatform\Core\Api\OperationAwareFormatsProviderInterface;
 use ApiPlatform\Core\Api\OperationMethodResolverInterface;
 use ApiPlatform\Core\Api\OperationType;
 use ApiPlatform\Core\Api\ResourceClassResolverInterface;
-use ApiPlatform\Core\Api\UrlGeneratorInterface;
 use ApiPlatform\Core\Bridge\Symfony\Routing\RouterOperationPathResolver;
 use ApiPlatform\Core\Documentation\Documentation;
 use ApiPlatform\Core\Exception\InvalidArgumentException;
@@ -48,35 +47,16 @@ use Symfony\Component\PropertyInfo\Type;
 use Symfony\Component\Routing\Route;
 use Symfony\Component\Routing\RouteCollection;
 use Symfony\Component\Routing\RouterInterface;
-use Symfony\Component\Serializer\NameConverter\AdvancedNameConverterInterface;
 use Symfony\Component\Serializer\NameConverter\NameConverterInterface;
 use Symfony\Component\Serializer\Normalizer\AbstractNormalizer;
 
 /**
  * @author Amrouche Hamza <hamza.simperfit@gmail.com>
  * @author Kévin Dunglas <dunglas@gmail.com>
+ * @author Anthony GRASSIOT <antograssiot@free.fr>
  */
-class DocumentationNormalizerTest extends TestCase
+class DocumentationNormalizerV3Test extends TestCase
 {
-    /**
-     * @group legacy
-     * @expectedDeprecation Passing an instance of ApiPlatform\Core\Api\UrlGeneratorInterface to ApiPlatform\Core\Swagger\Serializer\DocumentationNormalizer::__construct() is deprecated since version 2.1 and will be removed in 3.0.
-     */
-    public function testLegacyConstruct()
-    {
-        $normalizer = new DocumentationNormalizer(
-            $this->prophesize(ResourceMetadataFactoryInterface::class)->reveal(),
-            $this->prophesize(PropertyNameCollectionFactoryInterface::class)->reveal(),
-            $this->prophesize(PropertyMetadataFactoryInterface::class)->reveal(),
-            $this->prophesize(ResourceClassResolverInterface::class)->reveal(),
-            $this->prophesize(OperationMethodResolverInterface::class)->reveal(),
-            $this->prophesize(OperationPathResolverInterface::class)->reveal(),
-            $this->prophesize(UrlGeneratorInterface::class)->reveal()
-        );
-
-        $this->assertInstanceOf(DocumentationNormalizer::class, $normalizer);
-    }
-
     public function testNormalize()
     {
         $documentation = new Documentation(new ResourceNameCollection([Dummy::class]), 'Test API', 'This is a test API.', '1.2.3', ['jsonld' => ['application/ld+json']]);
@@ -111,12 +91,30 @@ class DocumentationNormalizerTest extends TestCase
             $propertyMetadataFactoryProphecy->reveal(),
             $resourceClassResolverProphecy->reveal(),
             $operationMethodResolverProphecy->reveal(),
-            $operationPathResolver
+            $operationPathResolver,
+            null,
+            null,
+            null,
+            false,
+            '',
+            '',
+            '',
+            '',
+            [],
+            [],
+            null,
+            true,
+            'page',
+            false,
+            'itemsPerPage',
+            null,
+            false,
+            'pagination',
+            ['spec_version' => 3]
         );
 
         $expected = [
-            'swagger' => '2.0',
-            'basePath' => '/app_dev.php/',
+            'openapi' => '3.0.2',
             'info' => [
                 'title' => 'Test API',
                 'description' => 'This is a test API.',
@@ -127,30 +125,37 @@ class DocumentationNormalizerTest extends TestCase
                     'get' => new \ArrayObject([
                         'tags' => ['Dummy'],
                         'operationId' => 'getDummyCollection',
-                        'produces' => ['application/ld+json'],
                         'summary' => 'Retrieves the collection of Dummy resources.',
                         'parameters' => [
                             [
                                 'name' => 'page',
                                 'in' => 'query',
                                 'required' => false,
-                                'type' => 'integer',
+                                'schema' => [
+                                    'type' => 'integer',
+                                ],
                                 'description' => 'The collection page number',
                             ],
                             [
                                 'name' => 'itemsPerPage',
                                 'in' => 'query',
                                 'required' => false,
-                                'type' => 'integer',
+                                'schema' => [
+                                    'type' => 'integer',
+                                ],
                                 'description' => 'The number of items per page',
                             ],
                         ],
                         'responses' => [
-                            200 => [
+                            '200' => [
                                 'description' => 'Dummy collection response',
-                                'schema' => [
-                                    'type' => 'array',
-                                    'items' => ['$ref' => '#/definitions/Dummy'],
+                                'content' => [
+                                    'application/ld+json' => [
+                                        'schema' => [
+                                            'type' => 'array',
+                                            'items' => ['$ref' => '#/components/schemas/Dummy'],
+                                        ],
+                                    ],
                                 ],
                             ],
                         ],
@@ -158,24 +163,26 @@ class DocumentationNormalizerTest extends TestCase
                     'post' => new \ArrayObject([
                         'tags' => ['Dummy'],
                         'operationId' => 'postDummyCollection',
-                        'consumes' => ['application/ld+json'],
-                        'produces' => ['application/ld+json'],
-                        'summary' => 'Creates a Dummy resource.',
-                        'parameters' => [
-                            [
-                                'name' => 'dummy',
-                                'in' => 'body',
-                                'description' => 'The new Dummy resource',
-                                'schema' => ['$ref' => '#/definitions/Dummy'],
+                        'requestBody' => [
+                            'content' => [
+                                'application/ld+json' => [
+                                    'schema' => ['$ref' => '#/components/schemas/Dummy'],
+                                ],
                             ],
+                            'description' => 'The new Dummy resource',
                         ],
+                        'summary' => 'Creates a Dummy resource.',
                         'responses' => [
-                            201 => [
+                            '201' => [
                                 'description' => 'Dummy resource created',
-                                'schema' => ['$ref' => '#/definitions/Dummy'],
+                                'content' => [
+                                    'application/ld+json' => [
+                                        'schema' => ['$ref' => '#/components/schemas/Dummy'],
+                                    ],
+                                ],
                             ],
-                            400 => ['description' => 'Invalid input'],
-                            404 => ['description' => 'Resource not found'],
+                            '400' => ['description' => 'Invalid input'],
+                            '404' => ['description' => 'Resource not found'],
                         ],
                     ]),
                 ],
@@ -183,51 +190,58 @@ class DocumentationNormalizerTest extends TestCase
                     'get' => new \ArrayObject([
                         'tags' => ['Dummy'],
                         'operationId' => 'getDummyItem',
-                        'produces' => ['application/ld+json'],
                         'summary' => 'Retrieves a Dummy resource.',
                         'parameters' => [
                             [
                                 'name' => 'id',
                                 'in' => 'path',
-                                'type' => 'string',
+                                'schema' => ['type' => 'string'],
                                 'required' => true,
                             ],
                         ],
                         'responses' => [
-                            200 => [
+                            '200' => [
                                 'description' => 'Dummy resource response',
-                                'schema' => ['$ref' => '#/definitions/Dummy'],
+                                'content' => [
+                                    'application/ld+json' => [
+                                        'schema' => ['$ref' => '#/components/schemas/Dummy'],
+                                    ],
+                                ],
                             ],
-                            404 => ['description' => 'Resource not found'],
+                            '404' => ['description' => 'Resource not found'],
                         ],
                     ]),
                     'put' => new \ArrayObject([
                         'tags' => ['Dummy'],
                         'operationId' => 'putDummyItem',
-                        'consumes' => ['application/ld+json'],
-                        'produces' => ['application/ld+json'],
+                        'requestBody' => [
+                            'content' => [
+                                'application/ld+json' => [
+                                    'schema' => ['$ref' => '#/components/schemas/Dummy'],
+                                ],
+                            ],
+                            'description' => 'The updated Dummy resource',
+                        ],
                         'summary' => 'Replaces the Dummy resource.',
                         'parameters' => [
                             [
                                 'name' => 'id',
                                 'in' => 'path',
-                                'type' => 'string',
+                                'schema' => ['type' => 'string'],
                                 'required' => true,
-                            ],
-                            [
-                                'name' => 'dummy',
-                                'in' => 'body',
-                                'description' => 'The updated Dummy resource',
-                                'schema' => ['$ref' => '#/definitions/Dummy'],
                             ],
                         ],
                         'responses' => [
-                            200 => [
+                            '200' => [
                                 'description' => 'Dummy resource updated',
-                                'schema' => ['$ref' => '#/definitions/Dummy'],
+                                'content' => [
+                                    'application/ld+json' => [
+                                        'schema' => ['$ref' => '#/components/schemas/Dummy'],
+                                    ],
+                                ],
                             ],
-                            400 => ['description' => 'Invalid input'],
-                            404 => ['description' => 'Resource not found'],
+                            '400' => ['description' => 'Invalid input'],
+                            '404' => ['description' => 'Resource not found'],
                         ],
                     ]),
                 ],
@@ -235,30 +249,33 @@ class DocumentationNormalizerTest extends TestCase
                     'get' => new \ArrayObject([
                         'tags' => ['Dummy'],
                         'operationId' => 'customDummyCollection',
-                        'produces' => ['application/ld+json'],
                         'summary' => 'Retrieves the collection of Dummy resources.',
                         'parameters' => [
                             [
                                 'name' => 'page',
                                 'in' => 'query',
                                 'required' => false,
-                                'type' => 'integer',
+                                'schema' => ['type' => 'integer'],
                                 'description' => 'The collection page number',
                             ],
                             [
                                 'name' => 'itemsPerPage',
                                 'in' => 'query',
                                 'required' => false,
-                                'type' => 'integer',
+                                'schema' => ['type' => 'integer'],
                                 'description' => 'The number of items per page',
                             ],
                         ],
                         'responses' => [
-                            200 => [
+                            '200' => [
                                 'description' => 'Dummy collection response',
-                                'schema' => [
-                                    'type' => 'array',
-                                    'items' => ['$ref' => '#/definitions/Dummy'],
+                                'content' => [
+                                    'application/ld+json' => [
+                                        'schema' => [
+                                            'type' => 'array',
+                                            'items' => ['$ref' => '#/components/schemas/Dummy'],
+                                        ],
+                                    ],
                                 ],
                             ],
                         ],
@@ -266,50 +283,55 @@ class DocumentationNormalizerTest extends TestCase
                     'post' => new \ArrayObject([
                         'tags' => ['Dummy'],
                         'operationId' => 'custom2DummyCollection',
-                        'produces' => ['application/ld+json'],
-                        'consumes' => ['application/ld+json'],
                         'summary' => 'Creates a Dummy resource.',
-                        'parameters' => [
-                            [
-                                'name' => 'dummy',
-                                'in' => 'body',
-                                'description' => 'The new Dummy resource',
-                                'schema' => ['$ref' => '#/definitions/Dummy'],
+                        'requestBody' => [
+                            'content' => [
+                                'application/ld+json' => [
+                                    'schema' => ['$ref' => '#/components/schemas/Dummy'],
+                                ],
                             ],
+                            'description' => 'The new Dummy resource',
                         ],
                         'responses' => [
-                            201 => [
+                            '201' => [
                                 'description' => 'Dummy resource created',
-                                'schema' => ['$ref' => '#/definitions/Dummy'],
+                                'content' => [
+                                    'application/ld+json' => [
+                                        'schema' => ['$ref' => '#/components/schemas/Dummy'],
+                                    ],
+                                ],
                             ],
-                            400 => ['description' => 'Invalid input'],
-                            404 => ['description' => 'Resource not found'],
+                            '400' => ['description' => 'Invalid input'],
+                            '404' => ['description' => 'Resource not found'],
                         ],
                     ]),
                 ],
             ]),
-            'definitions' => new \ArrayObject([
-                'Dummy' => new \ArrayObject([
-                    'type' => 'object',
-                    'description' => 'This is a dummy.',
-                    'externalDocs' => ['url' => 'http://schema.example.com/Dummy'],
-                    'properties' => [
-                        'id' => new \ArrayObject([
-                            'type' => 'integer',
-                            'description' => 'This is an id.',
-                            'readOnly' => true,
-                        ]),
-                        'name' => new \ArrayObject([
-                            'type' => 'string',
-                            'description' => 'This is a name.',
-                        ]),
-                        'description' => new \ArrayObject([
-                            'type' => 'string',
-                            'description' => 'This is an initializable but not writable property.',
-                        ]),
-                    ],
+            'components' => [
+                'schemas' => new \ArrayObject([
+                    'Dummy' => new \ArrayObject([
+                        'type' => 'object',
+                        'description' => 'This is a dummy.',
+                        'externalDocs' => ['url' => 'http://schema.example.com/Dummy'],
+                        'properties' => [
+                            'id' => new \ArrayObject([
+                                'type' => 'integer',
+                                'description' => 'This is an id.',
+                                'readOnly' => true,
+                            ]),
+                            'name' => new \ArrayObject([
+                                'type' => 'string',
+                                'description' => 'This is a name.',
+                            ]),
+                            'description' => new \ArrayObject([
+                                'type' => 'string',
+                                'description' => 'This is an initializable but not writable property.',
+                            ]),
+                        ],
+                    ]),
                 ]),
-            ]),
+            ],
+            'servers' => [['url' => '/app_dev.php/']],
         ];
 
         $this->assertEquals($expected, $normalizer->normalize($documentation, DocumentationNormalizer::FORMAT, ['base_url' => '/app_dev.php/']));
@@ -337,11 +359,7 @@ class DocumentationNormalizerTest extends TestCase
         $operationMethodResolverProphecy = $this->prophesize(OperationMethodResolverInterface::class);
         $operationMethodResolverProphecy->getItemOperationMethod(Dummy::class, 'get')->shouldBeCalled()->willReturn('GET');
 
-        $nameConverterProphecy = $this->prophesize(
-            interface_exists(AdvancedNameConverterInterface::class)
-                ? AdvancedNameConverterInterface::class
-                : NameConverterInterface::class
-        );
+        $nameConverterProphecy = $this->prophesize(NameConverterInterface::class);
         $nameConverterProphecy->normalize('name', Dummy::class, DocumentationNormalizer::FORMAT, [])->willReturn('name')->shouldBeCalled();
         $nameConverterProphecy->normalize('nameConverted', Dummy::class, DocumentationNormalizer::FORMAT, [])->willReturn('name_converted')->shouldBeCalled();
 
@@ -362,12 +380,21 @@ class DocumentationNormalizerTest extends TestCase
             'application',
             '/oauth/v2/token',
             '/oauth/v2/auth',
-            ['scope param']
+            ['scope param'],
+            [],
+            null,
+            true,
+            'page',
+            false,
+            'itemsPerPage',
+            null,
+            false,
+            'pagination',
+            ['spec_version' => 3]
         );
 
         $expected = [
-            'swagger' => '2.0',
-            'basePath' => '/',
+            'openapi' => '3.0.2',
             'info' => [
                 'title' => 'Dummy API',
                 'description' => 'This is a dummy API',
@@ -378,53 +405,59 @@ class DocumentationNormalizerTest extends TestCase
                     'get' => new \ArrayObject([
                         'tags' => ['Dummy'],
                         'operationId' => 'getDummyItem',
-                        'produces' => ['application/ld+json'],
                         'summary' => 'Retrieves a Dummy resource.',
                         'parameters' => [
                             [
                                 'name' => 'id',
                                 'in' => 'path',
-                                'type' => 'string',
+                                'schema' => ['type' => 'string'],
                                 'required' => true,
                             ],
                         ],
                         'responses' => [
-                            200 => [
+                            '200' => [
                                 'description' => 'Dummy resource response',
-                                'schema' => ['$ref' => '#/definitions/Dummy'],
+                                'content' => [
+                                    'application/ld+json' => [
+                                        'schema' => ['$ref' => '#/components/schemas/Dummy'],
+                                    ],
+                                ],
                             ],
-                            404 => ['description' => 'Resource not found'],
+                            '404' => ['description' => 'Resource not found'],
                         ],
                     ]),
                 ],
             ]),
-            'definitions' => new \ArrayObject([
-                'Dummy' => new \ArrayObject([
-                    'type' => 'object',
-                    'description' => 'This is a dummy.',
-                    'properties' => [
-                        'name' => new \ArrayObject([
-                            'type' => 'string',
-                            'description' => 'This is a name.',
-                        ]),
-                        'name_converted' => new \ArrayObject([
-                            'type' => 'string',
-                            'description' => 'This is a converted name.',
-                        ]),
-                    ],
+            'components' => [
+                'schemas' => new \ArrayObject([
+                    'Dummy' => new \ArrayObject([
+                        'type' => 'object',
+                        'description' => 'This is a dummy.',
+                        'properties' => [
+                            'name' => new \ArrayObject([
+                                'type' => 'string',
+                                'description' => 'This is a name.',
+                            ]),
+                            'name_converted' => new \ArrayObject([
+                                'type' => 'string',
+                                'description' => 'This is a converted name.',
+                            ]),
+                        ],
+                    ]),
                 ]),
-            ]),
-            'securityDefinitions' => [
-                'oauth' => [
-                    'type' => 'oauth2',
-                    'description' => 'OAuth client_credentials Grant',
-                    'flow' => 'application',
-                    'tokenUrl' => '/oauth/v2/token',
-                    'authorizationUrl' => '/oauth/v2/auth',
-                    'scopes' => ['scope param'],
+                'securitySchemes' => [
+                    'oauth' => [
+                        'type' => 'oauth2',
+                        'description' => 'OAuth client_credentials Grant',
+                        'flow' => 'application',
+                        'tokenUrl' => '/oauth/v2/token',
+                        'authorizationUrl' => '/oauth/v2/auth',
+                        'scopes' => ['scope param'],
+                    ],
                 ],
             ],
             'security' => [['oauth' => []]],
+            'servers' => [['url' => '/']],
         ];
 
         $this->assertEquals($expected, $normalizer->normalize($documentation));
@@ -480,12 +513,21 @@ class DocumentationNormalizerTest extends TestCase
             '',
             '',
             [],
-            $apiKeysConfiguration
+            $apiKeysConfiguration,
+            null,
+            true,
+            'page',
+            false,
+            'itemsPerPage',
+            null,
+            false,
+            'pagination',
+            ['spec_version' => 3]
         );
 
         $expected = [
-            'swagger' => '2.0',
-            'basePath' => '/app_dev.php/',
+            'openapi' => '3.0.2',
+            'servers' => [['url' => '/app_dev.php/']],
             'info' => [
                 'title' => 'Test API',
                 'description' => 'This is a test API.',
@@ -496,50 +538,55 @@ class DocumentationNormalizerTest extends TestCase
                     'get' => new \ArrayObject([
                         'tags' => ['Dummy'],
                         'operationId' => 'getDummyItem',
-                        'produces' => ['application/ld+json'],
                         'summary' => 'Retrieves a Dummy resource.',
                         'parameters' => [
                             [
                                 'name' => 'id',
                                 'in' => 'path',
-                                'type' => 'string',
+                                'schema' => ['type' => 'string'],
                                 'required' => true,
                             ],
                         ],
                         'responses' => [
-                            200 => [
+                            '200' => [
                                 'description' => 'Dummy resource response',
-                                'schema' => ['$ref' => '#/definitions/Dummy'],
+                                'content' => [
+                                    'application/ld+json' => [
+                                        'schema' => ['$ref' => '#/components/schemas/Dummy'],
+                                    ],
+                                ],
                             ],
-                            404 => ['description' => 'Resource not found'],
+                            '404' => ['description' => 'Resource not found'],
                         ],
                     ]),
                 ],
             ]),
-            'definitions' => new \ArrayObject([
-                'Dummy' => new \ArrayObject([
-                    'type' => 'object',
-                    'description' => 'This is a dummy.',
-                    'properties' => [
-                        'name' => new \ArrayObject([
-                            'type' => 'string',
-                            'description' => 'This is a name.',
-                        ]),
-                    ],
+            'components' => [
+                'schemas' => new \ArrayObject([
+                    'Dummy' => new \ArrayObject([
+                        'type' => 'object',
+                        'description' => 'This is a dummy.',
+                        'properties' => [
+                            'name' => new \ArrayObject([
+                                'type' => 'string',
+                                'description' => 'This is a name.',
+                            ]),
+                        ],
+                    ]),
                 ]),
-            ]),
-            'securityDefinitions' => [
-                'header' => [
-                    'type' => 'apiKey',
-                    'in' => 'header',
-                    'description' => 'Value for the Authorization header',
-                    'name' => 'Authorization',
-                ],
-                'query' => [
-                    'type' => 'apiKey',
-                    'in' => 'query',
-                    'description' => 'Value for the key query parameter',
-                    'name' => 'key',
+                'securitySchemes' => [
+                    'header' => [
+                        'type' => 'apiKey',
+                        'in' => 'header',
+                        'description' => 'Value for the Authorization header',
+                        'name' => 'Authorization',
+                    ],
+                    'query' => [
+                        'type' => 'apiKey',
+                        'in' => 'query',
+                        'description' => 'Value for the key query parameter',
+                        'name' => 'key',
+                    ],
                 ],
             ],
             'security' => [
@@ -604,12 +651,30 @@ class DocumentationNormalizerTest extends TestCase
             $propertyMetadataFactoryProphecy->reveal(),
             $resourceClassResolverProphecy->reveal(),
             $operationMethodResolverProphecy->reveal(),
-            $operationPathResolver
+            $operationPathResolver,
+            null,
+            null,
+            null,
+            false,
+            '',
+            '',
+            '',
+            '',
+            [],
+            [],
+            null,
+            true,
+            'page',
+            false,
+            'itemsPerPage',
+            null,
+            false,
+            'pagination',
+            ['spec_version' => 3]
         );
 
         $expected = [
-            'swagger' => '2.0',
-            'basePath' => '/',
+            'openapi' => '3.0.2',
             'info' => [
                 'title' => 'Test API',
                 'description' => 'This is a test API.',
@@ -622,23 +687,26 @@ class DocumentationNormalizerTest extends TestCase
                             'Dummy',
                         ],
                         'operationId' => 'getDummyCollection',
-                        'produces' => ['application/ld+json'],
                         'summary' => 'Retrieves the collection of Dummy resources.',
                         'parameters' => [
                             [
                                 'name' => 'page',
                                 'in' => 'query',
                                 'required' => false,
-                                'type' => 'integer',
+                                'schema' => ['type' => 'integer'],
                                 'description' => 'The collection page number',
                             ],
                         ],
                         'responses' => [
-                            200 => [
+                            '200' => [
                                 'description' => 'Dummy collection response',
-                                'schema' => [
-                                    'type' => 'array',
-                                    'items' => ['$ref' => '#/definitions/Dummy'],
+                                'content' => [
+                                    'application/ld+json' => [
+                                        'schema' => [
+                                            'type' => 'array',
+                                            'items' => ['$ref' => '#/components/schemas/Dummy'],
+                                        ],
+                                    ],
                                 ],
                             ],
                         ],
@@ -646,22 +714,26 @@ class DocumentationNormalizerTest extends TestCase
                     'post' => new \ArrayObject([
                         'tags' => ['Dummy'],
                         'operationId' => 'postDummyCollection',
-                        'consumes' => ['application/ld+json'],
-                        'produces' => ['application/ld+json'],
                         'summary' => 'Creates a Dummy resource.',
-                        'parameters' => [[
-                            'name' => 'dummy',
-                            'in' => 'body',
-                            'description' => 'The new Dummy resource',
-                            'schema' => ['$ref' => '#/definitions/Dummy'],
-                        ]],
-                        'responses' => [
-                            201 => [
-                                'description' => 'Dummy resource created',
-                                'schema' => ['$ref' => '#/definitions/Dummy'],
+                        'requestBody' => [
+                            'content' => [
+                                'application/ld+json' => [
+                                    'schema' => ['$ref' => '#/components/schemas/Dummy'],
+                                ],
                             ],
-                            400 => ['description' => 'Invalid input'],
-                            404 => ['description' => 'Resource not found'],
+                            'description' => 'The new Dummy resource',
+                        ],
+                        'responses' => [
+                            '201' => [
+                                'description' => 'Dummy resource created',
+                                'content' => [
+                                    'application/ld+json' => [
+                                        'schema' => ['$ref' => '#/components/schemas/Dummy'],
+                                    ],
+                                ],
+                            ],
+                            '400' => ['description' => 'Invalid input'],
+                            '404' => ['description' => 'Resource not found'],
                         ],
                     ]),
                 ],
@@ -669,83 +741,95 @@ class DocumentationNormalizerTest extends TestCase
                     'get' => new \ArrayObject([
                         'tags' => ['Dummy'],
                         'operationId' => 'getDummyItem',
-                        'produces' => ['application/ld+json'],
                         'summary' => 'Retrieves a Dummy resource.',
-                        'parameters' => [[
-                            'name' => 'id',
-                            'in' => 'path',
-                            'type' => 'string',
-                            'required' => true,
-                        ]],
-                        'responses' => [
-                            200 => [
-                                'description' => 'Dummy resource response',
-                                'schema' => ['$ref' => '#/definitions/Dummy'],
+                        'parameters' => [
+                            [
+                                'name' => 'id',
+                                'in' => 'path',
+                                'schema' => ['type' => 'string'],
+                                'required' => true,
                             ],
-                            404 => ['description' => 'Resource not found'],
+                        ],
+                        'responses' => [
+                            '200' => [
+                                'description' => 'Dummy resource response',
+                                'content' => [
+                                    'application/ld+json' => [
+                                        'schema' => ['$ref' => '#/components/schemas/Dummy'],
+                                    ],
+                                ],
+                            ],
+                            '404' => ['description' => 'Resource not found'],
                         ],
                     ]),
                     'put' => new \ArrayObject([
                         'tags' => ['Dummy'],
                         'operationId' => 'putDummyItem',
-                        'consumes' => ['application/ld+json'],
-                        'produces' => ['application/ld+json'],
+                        'requestBody' => [
+                            'content' => [
+                                'application/ld+json' => [
+                                    'schema' => ['$ref' => '#/components/schemas/Dummy'],
+                                ],
+                            ],
+                            'description' => 'The updated Dummy resource',
+                        ],
                         'summary' => 'Replaces the Dummy resource.',
                         'parameters' => [
                             [
                                 'name' => 'id',
                                 'in' => 'path',
-                                'type' => 'string',
+                                'schema' => ['type' => 'string'],
                                 'required' => true,
-                            ],
-                            [
-                                'name' => 'dummy',
-                                'in' => 'body',
-                                'description' => 'The updated Dummy resource',
-                                'schema' => ['$ref' => '#/definitions/Dummy'],
                             ],
                         ],
                         'responses' => [
-                            200 => [
+                            '200' => [
                                 'description' => 'Dummy resource updated',
-                                'schema' => ['$ref' => '#/definitions/'.$ref],
+                                'content' => [
+                                    'application/ld+json' => [
+                                        'schema' => ['$ref' => '#/components/schemas/'.$ref],
+                                    ],
+                                ],
                             ],
-                            400 => ['description' => 'Invalid input'],
-                            404 => ['description' => 'Resource not found'],
+                            '400' => ['description' => 'Invalid input'],
+                            '404' => ['description' => 'Resource not found'],
                         ],
                     ]),
                 ],
             ]),
-            'definitions' => new \ArrayObject([
-                'Dummy' => new \ArrayObject([
-                    'type' => 'object',
-                    'description' => 'This is a dummy.',
-                    'externalDocs' => ['url' => 'http://schema.example.com/Dummy'],
-                    'properties' => [
-                        'name' => new \ArrayObject([
-                            'type' => 'string',
-                            'description' => 'This is a name.',
-                        ]),
-                    ],
+            'components' => [
+                'schemas' => new \ArrayObject([
+                    'Dummy' => new \ArrayObject([
+                        'type' => 'object',
+                        'description' => 'This is a dummy.',
+                        'externalDocs' => ['url' => 'http://schema.example.com/Dummy'],
+                        'properties' => [
+                            'name' => new \ArrayObject([
+                                'type' => 'string',
+                                'description' => 'This is a name.',
+                            ]),
+                        ],
+                    ]),
+                    $ref => new \ArrayObject([
+                        'type' => 'object',
+                        'description' => 'This is a dummy.',
+                        'externalDocs' => ['url' => 'http://schema.example.com/Dummy'],
+                        'properties' => [
+                            'gerard' => new \ArrayObject([
+                                'type' => 'string',
+                                'description' => 'This is a gerard.',
+                            ]),
+                        ],
+                    ]),
                 ]),
-                $ref => new \ArrayObject([
-                    'type' => 'object',
-                    'description' => 'This is a dummy.',
-                    'externalDocs' => ['url' => 'http://schema.example.com/Dummy'],
-                    'properties' => [
-                        'gerard' => new \ArrayObject([
-                            'type' => 'string',
-                            'description' => 'This is a gerard.',
-                        ]),
-                    ],
-                ]),
-            ]),
+            ],
+            'servers' => [['url' => '/']],
         ];
 
         $this->assertEquals($expected, $normalizer->normalize($documentation));
     }
 
-    public function testNormalizeWithSwaggerDefinitionName()
+    public function testNormalizeWithOpenApiDefinitionName()
     {
         $documentation = new Documentation(new ResourceNameCollection([Dummy::class]), 'Test API', 'This is a test API.', '1.2.3', ['jsonld' => ['application/ld+json']]);
 
@@ -784,12 +868,30 @@ class DocumentationNormalizerTest extends TestCase
             $propertyMetadataFactoryProphecy->reveal(),
             $resourceClassResolverProphecy->reveal(),
             $operationMethodResolverProphecy->reveal(),
-            $operationPathResolver
+            $operationPathResolver,
+            null,
+            null,
+            null,
+            false,
+            '',
+            '',
+            '',
+            '',
+            [],
+            [],
+            null,
+            true,
+            'page',
+            false,
+            'itemsPerPage',
+            null,
+            false,
+            'pagination',
+            ['spec_version' => 3]
         );
 
         $expected = [
-            'swagger' => '2.0',
-            'basePath' => '/app_dev.php/',
+            'openapi' => '3.0.2',
             'info' => [
                 'title' => 'Test API',
                 'description' => 'This is a test API.',
@@ -800,40 +902,46 @@ class DocumentationNormalizerTest extends TestCase
                     'get' => new \ArrayObject([
                         'tags' => ['Dummy'],
                         'operationId' => 'getDummyItem',
-                        'produces' => ['application/ld+json'],
                         'summary' => 'Retrieves a Dummy resource.',
                         'parameters' => [
                             [
                                 'name' => 'id',
                                 'in' => 'path',
-                                'type' => 'string',
+                                'schema' => ['type' => 'string'],
                                 'required' => true,
                             ],
                         ],
                         'responses' => [
-                            200 => [
+                            '200' => [
                                 'description' => 'Dummy resource response',
-                                'schema' => ['$ref' => '#/definitions/Dummy-Read'],
+                                'content' => [
+                                    'application/ld+json' => [
+                                        'schema' => ['$ref' => '#/components/schemas/Dummy-Read'],
+                                    ],
+                                ],
                             ],
-                            404 => ['description' => 'Resource not found'],
+                            '404' => ['description' => 'Resource not found'],
                         ],
                     ]),
                 ],
             ]),
-            'definitions' => new \ArrayObject([
-                'Dummy-Read' => new \ArrayObject([
-                    'type' => 'object',
-                    'description' => 'This is a dummy.',
-                    'externalDocs' => ['url' => 'http://schema.example.com/Dummy'],
-                    'properties' => [
-                        'id' => new \ArrayObject([
-                            'type' => 'integer',
-                            'description' => 'This is an id.',
-                            'readOnly' => true,
-                        ]),
-                    ],
+            'components' => [
+                'schemas' => new \ArrayObject([
+                    'Dummy-Read' => new \ArrayObject([
+                        'type' => 'object',
+                        'description' => 'This is a dummy.',
+                        'externalDocs' => ['url' => 'http://schema.example.com/Dummy'],
+                        'properties' => [
+                            'id' => new \ArrayObject([
+                                'type' => 'integer',
+                                'description' => 'This is an id.',
+                                'readOnly' => true,
+                            ]),
+                        ],
+                    ]),
                 ]),
-            ]),
+            ],
+            'servers' => [['url' => '/app_dev.php/']],
         ];
 
         $this->assertEquals($expected, $normalizer->normalize($documentation, DocumentationNormalizer::FORMAT, ['base_url' => '/app_dev.php/']));
@@ -889,12 +997,30 @@ class DocumentationNormalizerTest extends TestCase
             $propertyMetadataFactoryProphecy->reveal(),
             $resourceClassResolverProphecy->reveal(),
             $operationMethodResolverProphecy->reveal(),
-            $operationPathResolver
+            $operationPathResolver,
+            null,
+            null,
+            null,
+            false,
+            '',
+            '',
+            '',
+            '',
+            [],
+            [],
+            null,
+            true,
+            'page',
+            false,
+            'itemsPerPage',
+            null,
+            false,
+            'pagination',
+            ['spec_version' => 3]
         );
 
         $expected = [
-            'swagger' => '2.0',
-            'basePath' => '/',
+            'openapi' => '3.0.2',
             'info' => [
                 'title' => 'Test API',
                 'description' => 'This is a test API.',
@@ -907,23 +1033,26 @@ class DocumentationNormalizerTest extends TestCase
                             'Dummy',
                         ],
                         'operationId' => 'getDummyCollection',
-                        'produces' => ['application/ld+json'],
                         'summary' => 'Retrieves the collection of Dummy resources.',
                         'parameters' => [
                             [
                                 'name' => 'page',
                                 'in' => 'query',
                                 'required' => false,
-                                'type' => 'integer',
+                                'schema' => ['type' => 'integer'],
                                 'description' => 'The collection page number',
                             ],
                         ],
                         'responses' => [
-                            200 => [
+                            '200' => [
                                 'description' => 'Dummy collection response',
-                                'schema' => [
-                                    'type' => 'array',
-                                    'items' => ['$ref' => '#/definitions/Dummy'],
+                                'content' => [
+                                    'application/ld+json' => [
+                                        'schema' => [
+                                            'type' => 'array',
+                                            'items' => ['$ref' => '#/components/schemas/Dummy'],
+                                        ],
+                                    ],
                                 ],
                             ],
                         ],
@@ -931,22 +1060,26 @@ class DocumentationNormalizerTest extends TestCase
                     'post' => new \ArrayObject([
                         'tags' => ['Dummy'],
                         'operationId' => 'postDummyCollection',
-                        'consumes' => ['application/ld+json'],
-                        'produces' => ['application/ld+json'],
-                        'summary' => 'Creates a Dummy resource.',
-                        'parameters' => [[
-                            'name' => 'dummy',
-                            'in' => 'body',
-                            'description' => 'The new Dummy resource',
-                            'schema' => ['$ref' => '#/definitions/Dummy'],
-                        ]],
-                        'responses' => [
-                            201 => [
-                                'description' => 'Dummy resource created',
-                                'schema' => ['$ref' => '#/definitions/Dummy'],
+                        'requestBody' => [
+                            'content' => [
+                                'application/ld+json' => [
+                                    'schema' => ['$ref' => '#/components/schemas/Dummy'],
+                                ],
                             ],
-                            400 => ['description' => 'Invalid input'],
-                            404 => ['description' => 'Resource not found'],
+                            'description' => 'The new Dummy resource',
+                        ],
+                        'summary' => 'Creates a Dummy resource.',
+                        'responses' => [
+                            '201' => [
+                                'description' => 'Dummy resource created',
+                                'content' => [
+                                    'application/ld+json' => [
+                                        'schema' => ['$ref' => '#/components/schemas/Dummy'],
+                                    ],
+                                ],
+                            ],
+                            '400' => ['description' => 'Invalid input'],
+                            '404' => ['description' => 'Resource not found'],
                         ],
                     ]),
                 ],
@@ -954,77 +1087,87 @@ class DocumentationNormalizerTest extends TestCase
                     'get' => new \ArrayObject([
                         'tags' => ['Dummy'],
                         'operationId' => 'getDummyItem',
-                        'produces' => ['application/ld+json'],
                         'summary' => 'Retrieves a Dummy resource.',
                         'parameters' => [[
                             'name' => 'id',
                             'in' => 'path',
-                            'type' => 'string',
+                            'schema' => ['type' => 'string'],
                             'required' => true,
                         ]],
                         'responses' => [
-                            200 => [
+                            '200' => [
                                 'description' => 'Dummy resource response',
-                                'schema' => ['$ref' => '#/definitions/Dummy'],
+                                'content' => [
+                                    'application/ld+json' => [
+                                        'schema' => ['$ref' => '#/components/schemas/Dummy'],
+                                    ],
+                                ],
                             ],
-                            404 => ['description' => 'Resource not found'],
+                            '404' => ['description' => 'Resource not found'],
                         ],
                     ]),
                     'put' => new \ArrayObject([
                         'tags' => ['Dummy'],
                         'operationId' => 'putDummyItem',
-                        'consumes' => ['application/ld+json'],
-                        'produces' => ['application/ld+json'],
+                        'requestBody' => [
+                            'content' => [
+                                'application/ld+json' => [
+                                    'schema' => ['$ref' => '#/components/schemas/Dummy-dummy'],
+                                ],
+                            ],
+                            'description' => 'The updated Dummy resource',
+                        ],
                         'summary' => 'Replaces the Dummy resource.',
                         'parameters' => [
                             [
                                 'name' => 'id',
                                 'in' => 'path',
-                                'type' => 'string',
+                                'schema' => ['type' => 'string'],
                                 'required' => true,
-                            ],
-                            [
-                                'name' => 'dummy',
-                                'in' => 'body',
-                                'description' => 'The updated Dummy resource',
-                                'schema' => ['$ref' => '#/definitions/Dummy-dummy'],
                             ],
                         ],
                         'responses' => [
-                            200 => [
+                            '200' => [
                                 'description' => 'Dummy resource updated',
-                                'schema' => ['$ref' => '#/definitions/Dummy'],
+                                'content' => [
+                                    'application/ld+json' => [
+                                        'schema' => ['$ref' => '#/components/schemas/Dummy'],
+                                    ],
+                                ],
                             ],
-                            400 => ['description' => 'Invalid input'],
-                            404 => ['description' => 'Resource not found'],
+                            '400' => ['description' => 'Invalid input'],
+                            '404' => ['description' => 'Resource not found'],
                         ],
                     ]),
                 ],
             ]),
-            'definitions' => new \ArrayObject([
-                'Dummy' => new \ArrayObject([
-                    'type' => 'object',
-                    'description' => 'This is a dummy.',
-                    'externalDocs' => ['url' => 'http://schema.example.com/Dummy'],
-                    'properties' => [
-                        'name' => new \ArrayObject([
-                            'type' => 'string',
-                            'description' => 'This is a name.',
-                        ]),
-                    ],
+            'components' => [
+                'schemas' => new \ArrayObject([
+                    'Dummy' => new \ArrayObject([
+                        'type' => 'object',
+                        'description' => 'This is a dummy.',
+                        'externalDocs' => ['url' => 'http://schema.example.com/Dummy'],
+                        'properties' => [
+                            'name' => new \ArrayObject([
+                                'type' => 'string',
+                                'description' => 'This is a name.',
+                            ]),
+                        ],
+                    ]),
+                    'Dummy-dummy' => new \ArrayObject([
+                        'type' => 'object',
+                        'description' => 'This is a dummy.',
+                        'externalDocs' => ['url' => 'http://schema.example.com/Dummy'],
+                        'properties' => [
+                            'gerard' => new \ArrayObject([
+                                'type' => 'string',
+                                'description' => 'This is a gerard.',
+                            ]),
+                        ],
+                    ]),
                 ]),
-                'Dummy-dummy' => new \ArrayObject([
-                    'type' => 'object',
-                    'description' => 'This is a dummy.',
-                    'externalDocs' => ['url' => 'http://schema.example.com/Dummy'],
-                    'properties' => [
-                        'gerard' => new \ArrayObject([
-                            'type' => 'string',
-                            'description' => 'This is a gerard.',
-                        ]),
-                    ],
-                ]),
-            ]),
+            ],
+            'servers' => [['url' => '/']],
         ];
 
         $this->assertEquals($expected, $normalizer->normalize($documentation));
@@ -1083,12 +1226,30 @@ class DocumentationNormalizerTest extends TestCase
             $propertyMetadataFactoryProphecy->reveal(),
             $resourceClassResolverProphecy->reveal(),
             $operationMethodResolverProphecy->reveal(),
-            $operationPathResolver
+            $operationPathResolver,
+            null,
+            null,
+            null,
+            false,
+            '',
+            '',
+            '',
+            '',
+            [],
+            [],
+            null,
+            true,
+            'page',
+            false,
+            'itemsPerPage',
+            null,
+            false,
+            'pagination',
+            ['spec_version' => 3]
         );
 
         $expected = [
-            'swagger' => '2.0',
-            'basePath' => '/',
+            'openapi' => '3.0.2',
             'info' => [
                 'title' => 'Test API',
                 'description' => 'This is a test API.',
@@ -1101,23 +1262,26 @@ class DocumentationNormalizerTest extends TestCase
                             'Dummy',
                         ],
                         'operationId' => 'getDummyCollection',
-                        'produces' => ['application/ld+json'],
                         'summary' => 'Retrieves the collection of Dummy resources.',
                         'parameters' => [
                             [
                                 'name' => 'page',
                                 'in' => 'query',
                                 'required' => false,
-                                'type' => 'integer',
+                                'schema' => ['type' => 'integer'],
                                 'description' => 'The collection page number',
                             ],
                         ],
                         'responses' => [
-                            200 => [
+                            '200' => [
                                 'description' => 'Dummy collection response',
-                                'schema' => [
-                                    'type' => 'array',
-                                    'items' => ['$ref' => '#/definitions/Dummy'],
+                                'content' => [
+                                    'application/ld+json' => [
+                                        'schema' => [
+                                            'type' => 'array',
+                                            'items' => ['$ref' => '#/components/schemas/Dummy'],
+                                        ],
+                                    ],
                                 ],
                             ],
                         ],
@@ -1125,22 +1289,26 @@ class DocumentationNormalizerTest extends TestCase
                     'post' => new \ArrayObject([
                         'tags' => ['Dummy'],
                         'operationId' => 'postDummyCollection',
-                        'consumes' => ['application/ld+json'],
-                        'produces' => ['application/ld+json'],
-                        'summary' => 'Creates a Dummy resource.',
-                        'parameters' => [[
-                            'name' => 'dummy',
-                            'in' => 'body',
-                            'description' => 'The new Dummy resource',
-                            'schema' => ['$ref' => '#/definitions/Dummy'],
-                        ]],
-                        'responses' => [
-                            201 => [
-                                'description' => 'Dummy resource created',
-                                'schema' => ['$ref' => '#/definitions/Dummy'],
+                        'requestBody' => [
+                            'content' => [
+                                'application/ld+json' => [
+                                    'schema' => ['$ref' => '#/components/schemas/Dummy'],
+                                ],
                             ],
-                            400 => ['description' => 'Invalid input'],
-                            404 => ['description' => 'Resource not found'],
+                            'description' => 'The new Dummy resource',
+                        ],
+                        'summary' => 'Creates a Dummy resource.',
+                        'responses' => [
+                            '201' => [
+                                'description' => 'Dummy resource created',
+                                'content' => [
+                                    'application/ld+json' => [
+                                        'schema' => ['$ref' => '#/components/schemas/Dummy'],
+                                    ],
+                                ],
+                            ],
+                            '400' => ['description' => 'Invalid input'],
+                            '404' => ['description' => 'Resource not found'],
                         ],
                     ]),
                 ],
@@ -1148,77 +1316,87 @@ class DocumentationNormalizerTest extends TestCase
                     'get' => new \ArrayObject([
                         'tags' => ['Dummy'],
                         'operationId' => 'getDummyItem',
-                        'produces' => ['application/ld+json'],
                         'summary' => 'Retrieves a Dummy resource.',
                         'parameters' => [[
                             'name' => 'id',
                             'in' => 'path',
-                            'type' => 'string',
+                            'schema' => ['type' => 'string'],
                             'required' => true,
                         ]],
                         'responses' => [
-                            200 => [
+                            '200' => [
                                 'description' => 'Dummy resource response',
-                                'schema' => ['$ref' => '#/definitions/Dummy'],
+                                'content' => [
+                                    'application/ld+json' => [
+                                        'schema' => ['$ref' => '#/components/schemas/Dummy'],
+                                    ],
+                                ],
                             ],
-                            404 => ['description' => 'Resource not found'],
+                            '404' => ['description' => 'Resource not found'],
                         ],
                     ]),
                     'put' => new \ArrayObject([
                         'tags' => ['Dummy'],
                         'operationId' => 'putDummyItem',
-                        'consumes' => ['application/ld+json'],
-                        'produces' => ['application/ld+json'],
+                        'requestBody' => [
+                            'content' => [
+                                'application/ld+json' => [
+                                    'schema' => ['$ref' => '#/components/schemas/Dummy-dummy'],
+                                ],
+                            ],
+                            'description' => 'The updated Dummy resource',
+                        ],
                         'summary' => 'Replaces the Dummy resource.',
                         'parameters' => [
                             [
                                 'name' => 'id',
                                 'in' => 'path',
-                                'type' => 'string',
+                                'schema' => ['type' => 'string'],
                                 'required' => true,
-                            ],
-                            [
-                                'name' => 'dummy',
-                                'in' => 'body',
-                                'description' => 'The updated Dummy resource',
-                                'schema' => ['$ref' => '#/definitions/Dummy-dummy'],
                             ],
                         ],
                         'responses' => [
-                            200 => [
+                            '200' => [
                                 'description' => 'Dummy resource updated',
-                                'schema' => ['$ref' => '#/definitions/Dummy-dummy'],
+                                'content' => [
+                                    'application/ld+json' => [
+                                        'schema' => ['$ref' => '#/components/schemas/Dummy-dummy'],
+                                    ],
+                                ],
                             ],
-                            400 => ['description' => 'Invalid input'],
-                            404 => ['description' => 'Resource not found'],
+                            '400' => ['description' => 'Invalid input'],
+                            '404' => ['description' => 'Resource not found'],
                         ],
                     ]),
                 ],
             ]),
-            'definitions' => new \ArrayObject([
-                'Dummy' => new \ArrayObject([
-                    'type' => 'object',
-                    'description' => 'This is a dummy.',
-                    'externalDocs' => ['url' => 'http://schema.example.com/Dummy'],
-                    'properties' => [
-                        'name' => new \ArrayObject([
-                            'type' => 'string',
-                            'description' => 'This is a name.',
-                        ]),
-                    ],
+            'components' => [
+                'schemas' => new \ArrayObject([
+                    'Dummy' => new \ArrayObject([
+                        'type' => 'object',
+                        'description' => 'This is a dummy.',
+                        'externalDocs' => ['url' => 'http://schema.example.com/Dummy'],
+                        'properties' => [
+                            'name' => new \ArrayObject([
+                                'type' => 'string',
+                                'description' => 'This is a name.',
+                            ]),
+                        ],
+                    ]),
+                    'Dummy-dummy' => new \ArrayObject([
+                        'type' => 'object',
+                        'description' => 'This is a dummy.',
+                        'externalDocs' => ['url' => 'http://schema.example.com/Dummy'],
+                        'properties' => [
+                            'gerard' => new \ArrayObject([
+                                'type' => 'string',
+                                'description' => 'This is a gerard.',
+                            ]),
+                        ],
+                    ]),
                 ]),
-                'Dummy-dummy' => new \ArrayObject([
-                    'type' => 'object',
-                    'description' => 'This is a dummy.',
-                    'externalDocs' => ['url' => 'http://schema.example.com/Dummy'],
-                    'properties' => [
-                        'gerard' => new \ArrayObject([
-                            'type' => 'string',
-                            'description' => 'This is a gerard.',
-                        ]),
-                    ],
-                ]),
-            ]),
+            ],
+            'servers' => [['url' => '/']],
         ];
 
         $this->assertEquals($expected, $normalizer->normalize($documentation));
@@ -1233,7 +1411,7 @@ class DocumentationNormalizerTest extends TestCase
                 'type' => 'string',
                 'required' => true,
                 'strategy' => 'exact',
-                'swagger' => ['x-foo' => 'bar'],
+                'openapi' => ['x-foo' => 'bar'],
             ]]),
             'f2' => new DummyFilter(['ha' => [
                 'property' => 'foo',
@@ -1272,7 +1450,7 @@ class DocumentationNormalizerTest extends TestCase
                 'type' => 'string',
                 'required' => true,
                 'strategy' => 'exact',
-                'swagger' => ['x-foo' => 'bar'],
+                'openapi' => ['x-foo' => 'bar'],
             ]]),
             'f2' => new DummyFilter(['ha' => [
                 'property' => 'foo',
@@ -1303,7 +1481,24 @@ class DocumentationNormalizerTest extends TestCase
             $this->prophesize(OperationMethodResolverInterface::class)->reveal(),
             $this->prophesize(OperationPathResolverInterface::class)->reveal(),
             null,
-            new \ArrayObject()
+            new \ArrayObject(),
+            null,
+            false,
+            '',
+            '',
+            '',
+            '',
+            [],
+            [],
+            null,
+            true,
+            'page',
+            false,
+            'itemsPerPage',
+            null,
+            false,
+            'pagination',
+            ['spec_version' => 3]
         );
     }
 
@@ -1322,7 +1517,26 @@ class DocumentationNormalizerTest extends TestCase
             $propertyMetadataFactoryProphecy->reveal(),
             $resourceClassResolverProphecy->reveal(),
             $operationMethodResolverProphecy->reveal(),
-            $operationPathResolver
+            $operationPathResolver,
+            null,
+            null,
+            null,
+            false,
+            '',
+            '',
+            '',
+            '',
+            [],
+            [],
+            null,
+            true,
+            'page',
+            false,
+            'itemsPerPage',
+            null,
+            false,
+            'pagination',
+            ['spec_version' => 3]
         );
 
         $documentation = new Documentation(new ResourceNameCollection([Dummy::class]), 'Test API', 'This is a test API.', '1.2.3', ['jsonld' => ['application/ld+json']]);
@@ -1368,17 +1582,36 @@ class DocumentationNormalizerTest extends TestCase
             $propertyMetadataFactoryProphecy->reveal(),
             $resourceClassResolverProphecy->reveal(),
             $operationMethodResolverProphecy->reveal(),
-            $operationPathResolver
+            $operationPathResolver,
+            null,
+            null,
+            null,
+            false,
+            '',
+            '',
+            '',
+            '',
+            [],
+            [],
+            null,
+            true,
+            'page',
+            false,
+            'itemsPerPage',
+            null,
+            false,
+            'pagination',
+            ['spec_version' => 3]
         );
 
         $expected = [
-            'swagger' => '2.0',
-            'basePath' => '/',
+            'openapi' => '3.0.2',
             'info' => [
                 'title' => '',
                 'version' => '0.0.0',
             ],
             'paths' => new \ArrayObject([]),
+            'servers' => [['url' => '/']],
         ];
 
         $this->assertEquals($expected, $normalizer->normalize($documentation));
@@ -1417,12 +1650,30 @@ class DocumentationNormalizerTest extends TestCase
             $propertyMetadataFactoryProphecy->reveal(),
             $resourceClassResolverProphecy->reveal(),
             $operationMethodResolverProphecy->reveal(),
-            $operationPathResolver
+            $operationPathResolver,
+            null,
+            null,
+            null,
+            false,
+            '',
+            '',
+            '',
+            '',
+            [],
+            [],
+            null,
+            true,
+            'page',
+            false,
+            'itemsPerPage',
+            null,
+            false,
+            'pagination',
+            ['spec_version' => 3]
         );
 
         $expected = [
-            'swagger' => '2.0',
-            'basePath' => '/',
+            'openapi' => '3.0.2',
             'info' => [
                 'title' => '',
                 'version' => '0.0.0',
@@ -1435,6 +1686,7 @@ class DocumentationNormalizerTest extends TestCase
                     ]),
                 ],
             ]),
+            'servers' => [['url' => '/']],
         ];
 
         $this->assertEquals($expected, $normalizer->normalize($documentation));
@@ -1509,12 +1761,30 @@ class DocumentationNormalizerTest extends TestCase
             $propertyMetadataFactoryProphecy->reveal(),
             $resourceClassResolverProphecy->reveal(),
             $operationMethodResolverProphecy->reveal(),
-            $operationPathResolver
+            $operationPathResolver,
+            null,
+            null,
+            null,
+            false,
+            '',
+            '',
+            '',
+            '',
+            [],
+            [],
+            null,
+            true,
+            'page',
+            false,
+            'itemsPerPage',
+            null,
+            false,
+            'pagination',
+            ['spec_version' => 3]
         );
 
         $expected = [
-            'swagger' => '2.0',
-            'basePath' => '/',
+            'openapi' => '3.0.2',
             'info' => [
                 'title' => 'Test API',
                 'description' => 'This is a test API.',
@@ -1527,23 +1797,26 @@ class DocumentationNormalizerTest extends TestCase
                             'Dummy',
                         ],
                         'operationId' => 'getDummyCollection',
-                        'produces' => ['application/ld+json'],
                         'summary' => 'Retrieves the collection of Dummy resources.',
                         'parameters' => [
                             [
                                 'name' => 'page',
                                 'in' => 'query',
                                 'required' => false,
-                                'type' => 'integer',
+                                'schema' => ['type' => 'integer'],
                                 'description' => 'The collection page number',
                             ],
                         ],
                         'responses' => [
-                            200 => [
+                            '200' => [
                                 'description' => 'Dummy collection response',
-                                'schema' => [
-                                    'type' => 'array',
-                                    'items' => ['$ref' => '#/definitions/Dummy'],
+                                'content' => [
+                                    'application/ld+json' => [
+                                        'schema' => [
+                                            'type' => 'array',
+                                            'items' => ['$ref' => '#/components/schemas/Dummy'],
+                                        ],
+                                    ],
                                 ],
                             ],
                         ],
@@ -1551,22 +1824,26 @@ class DocumentationNormalizerTest extends TestCase
                     'post' => new \ArrayObject([
                         'tags' => ['Dummy'],
                         'operationId' => 'postDummyCollection',
-                        'consumes' => ['application/ld+json'],
-                        'produces' => ['application/ld+json'],
-                        'summary' => 'Creates a Dummy resource.',
-                        'parameters' => [[
-                            'name' => 'dummy',
-                            'in' => 'body',
-                            'description' => 'The new Dummy resource',
-                            'schema' => ['$ref' => '#/definitions/Dummy'],
-                        ]],
-                        'responses' => [
-                            201 => [
-                                'description' => 'Dummy resource created',
-                                'schema' => ['$ref' => '#/definitions/Dummy'],
+                        'requestBody' => [
+                            'content' => [
+                                'application/ld+json' => [
+                                    'schema' => ['$ref' => '#/components/schemas/Dummy'],
+                                ],
                             ],
-                            400 => ['description' => 'Invalid input'],
-                            404 => ['description' => 'Resource not found'],
+                            'description' => 'The new Dummy resource',
+                        ],
+                        'summary' => 'Creates a Dummy resource.',
+                        'responses' => [
+                            '201' => [
+                                'description' => 'Dummy resource created',
+                                'content' => [
+                                    'application/ld+json' => [
+                                        'schema' => ['$ref' => '#/components/schemas/Dummy'],
+                                    ],
+                                ],
+                            ],
+                            '400' => ['description' => 'Invalid input'],
+                            '404' => ['description' => 'Resource not found'],
                         ],
                     ]),
                 ],
@@ -1574,92 +1851,102 @@ class DocumentationNormalizerTest extends TestCase
                     'get' => new \ArrayObject([
                         'tags' => ['Dummy'],
                         'operationId' => 'getDummyItem',
-                        'produces' => ['application/ld+json'],
                         'summary' => 'Retrieves a Dummy resource.',
                         'parameters' => [[
                             'name' => 'id',
                             'in' => 'path',
-                            'type' => 'string',
+                            'schema' => ['type' => 'string'],
                             'required' => true,
                         ]],
                         'responses' => [
-                            200 => [
+                            '200' => [
                                 'description' => 'Dummy resource response',
-                                'schema' => ['$ref' => '#/definitions/Dummy'],
+                                'content' => [
+                                    'application/ld+json' => [
+                                        'schema' => ['$ref' => '#/components/schemas/Dummy'],
+                                    ],
+                                ],
                             ],
-                            404 => ['description' => 'Resource not found'],
+                            '404' => ['description' => 'Resource not found'],
                         ],
                     ]),
                     'put' => new \ArrayObject([
                         'tags' => ['Dummy'],
                         'operationId' => 'putDummyItem',
-                        'consumes' => ['application/ld+json'],
-                        'produces' => ['application/ld+json'],
+                        'requestBody' => [
+                            'content' => [
+                                'application/ld+json' => [
+                                    'schema' => ['$ref' => '#/components/schemas/Dummy'],
+                                ],
+                            ],
+                            'description' => 'The updated Dummy resource',
+                        ],
                         'summary' => 'Replaces the Dummy resource.',
                         'parameters' => [
                             [
                                 'name' => 'id',
                                 'in' => 'path',
-                                'type' => 'string',
+                                'schema' => ['type' => 'string'],
                                 'required' => true,
-                            ],
-                            [
-                                'name' => 'dummy',
-                                'in' => 'body',
-                                'description' => 'The updated Dummy resource',
-                                'schema' => ['$ref' => '#/definitions/Dummy'],
                             ],
                         ],
                         'responses' => [
-                            200 => [
+                            '200' => [
                                 'description' => 'Dummy resource updated',
-                                'schema' => ['$ref' => '#/definitions/'.$ref],
+                                'content' => [
+                                    'application/ld+json' => [
+                                        'schema' => ['$ref' => '#/components/schemas/'.$ref],
+                                    ],
+                                ],
                             ],
-                            400 => ['description' => 'Invalid input'],
-                            404 => ['description' => 'Resource not found'],
+                            '400' => ['description' => 'Invalid input'],
+                            '404' => ['description' => 'Resource not found'],
                         ],
                     ]),
                 ],
             ]),
-            'definitions' => new \ArrayObject([
-                'Dummy' => new \ArrayObject([
-                    'type' => 'object',
-                    'description' => 'This is a dummy.',
-                    'externalDocs' => ['url' => 'http://schema.example.com/Dummy'],
-                    'properties' => [
-                        'name' => new \ArrayObject([
-                            'type' => 'string',
-                            'description' => 'This is a name.',
-                        ]),
-                    ],
+            'components' => [
+                'schemas' => new \ArrayObject([
+                    'Dummy' => new \ArrayObject([
+                        'type' => 'object',
+                        'description' => 'This is a dummy.',
+                        'externalDocs' => ['url' => 'http://schema.example.com/Dummy'],
+                        'properties' => [
+                            'name' => new \ArrayObject([
+                                'type' => 'string',
+                                'description' => 'This is a name.',
+                            ]),
+                        ],
+                    ]),
+                    $ref => new \ArrayObject([
+                        'type' => 'object',
+                        'description' => 'This is a dummy.',
+                        'externalDocs' => ['url' => 'http://schema.example.com/Dummy'],
+                        'properties' => [
+                            'name' => new \ArrayObject([
+                                'type' => 'string',
+                                'description' => 'This is a name.',
+                            ]),
+                            'relatedDummy' => new \ArrayObject([
+                                'description' => 'This is a related dummy \o/.',
+                                '$ref' => '#/components/schemas/'.$relatedDummyRef,
+                            ]),
+                        ],
+                    ]),
+                    $relatedDummyRef => new \ArrayObject([
+                        'type' => 'object',
+                        'description' => 'This is a related dummy.',
+                        'externalDocs' => ['url' => 'http://schema.example.com/RelatedDummy'],
+                        'properties' => [
+                            'name' => new \ArrayObject([
+                                'type' => 'string',
+                                'description' => 'This is a name.',
+                            ]),
+                        ],
+                    ]),
                 ]),
-                $ref => new \ArrayObject([
-                    'type' => 'object',
-                    'description' => 'This is a dummy.',
-                    'externalDocs' => ['url' => 'http://schema.example.com/Dummy'],
-                    'properties' => [
-                        'name' => new \ArrayObject([
-                            'type' => 'string',
-                            'description' => 'This is a name.',
-                        ]),
-                        'relatedDummy' => new \ArrayObject([
-                            'description' => 'This is a related dummy \o/.',
-                            '$ref' => '#/definitions/'.$relatedDummyRef,
-                        ]),
-                    ],
-                ]),
-                $relatedDummyRef => new \ArrayObject([
-                    'type' => 'object',
-                    'description' => 'This is a related dummy.',
-                    'externalDocs' => ['url' => 'http://schema.example.com/RelatedDummy'],
-                    'properties' => [
-                        'name' => new \ArrayObject([
-                            'type' => 'string',
-                            'description' => 'This is a name.',
-                        ]),
-                    ],
-                ]),
-            ]),
+            ],
+            'servers' => [['url' => '/']],
         ];
 
         $this->assertEquals($expected, $normalizer->normalize($documentation));
@@ -1702,12 +1989,28 @@ class DocumentationNormalizerTest extends TestCase
             $operationMethodResolverProphecy->reveal(),
             $operationPathResolver,
             null,
-            $filterLocator
+            $filterLocator,
+            null,
+            false,
+            '',
+            '',
+            '',
+            '',
+            [],
+            [],
+            null,
+            true,
+            'page',
+            false,
+            'itemsPerPage',
+            null,
+            false,
+            'pagination',
+            ['spec_version' => 3]
         );
 
         $expected = [
-            'swagger' => '2.0',
-            'basePath' => '/',
+            'openapi' => '3.0.2',
             'info' => [
                 'title' => '',
                 'version' => '0.0.0',
@@ -1717,14 +2020,17 @@ class DocumentationNormalizerTest extends TestCase
                     'get' => new \ArrayObject([
                         'tags' => ['Dummy'],
                         'operationId' => 'getDummyCollection',
-                        'produces' => ['application/ld+json'],
                         'summary' => 'Retrieves the collection of Dummy resources.',
                         'responses' => [
-                            200 => [
+                            '200' => [
                                 'description' => 'Dummy collection response',
-                                'schema' => [
-                                    'type' => 'array',
-                                    'items' => ['$ref' => '#/definitions/Dummy'],
+                                'content' => [
+                                    'application/ld+json' => [
+                                        'schema' => [
+                                            'type' => 'array',
+                                            'items' => ['$ref' => '#/components/schemas/Dummy'],
+                                        ],
+                                    ],
                                 ],
                             ],
                         ],
@@ -1734,47 +2040,53 @@ class DocumentationNormalizerTest extends TestCase
                                 'name' => 'name',
                                 'in' => 'query',
                                 'required' => true,
-                                'type' => 'string',
+                                'schema' => ['type' => 'string'],
                             ],
                             [
                                 'name' => 'ha',
                                 'in' => 'query',
                                 'required' => false,
-                                'type' => 'integer',
+                                'schema' => ['type' => 'integer'],
                             ],
                             [
                                 'name' => 'toto',
                                 'in' => 'query',
                                 'required' => true,
-                                'type' => 'array',
-                                'items' => [
-                                    'type' => 'string',
+                                'schema' => [
+                                    'type' => 'array',
+                                    'items' => [
+                                        'type' => 'string',
+                                    ],
                                 ],
-                                'collectionFormat' => 'csv',
+                                'style' => 'deepObject',
+                                'explode' => true,
                             ],
                             [
                                 'name' => 'page',
                                 'in' => 'query',
                                 'required' => false,
-                                'type' => 'integer',
+                                'schema' => ['type' => 'integer'],
                                 'description' => 'The collection page number',
                             ],
                         ],
                     ]),
                 ],
             ]),
-            'definitions' => new \ArrayObject([
-                'Dummy' => new \ArrayObject([
-                    'type' => 'object',
-                    'description' => 'This is a dummy.',
-                    'properties' => [
-                        'name' => new \ArrayObject([
-                            'description' => 'This is a name.',
-                            'type' => 'string',
-                        ]),
-                    ],
+            'components' => [
+                'schemas' => new \ArrayObject([
+                    'Dummy' => new \ArrayObject([
+                        'type' => 'object',
+                        'description' => 'This is a dummy.',
+                        'properties' => [
+                            'name' => new \ArrayObject([
+                                'description' => 'This is a name.',
+                                'type' => 'string',
+                            ]),
+                        ],
+                    ]),
                 ]),
-            ]),
+            ],
+            'servers' => [['url' => '/']],
         ];
 
         $this->assertEquals($expected, $normalizer->normalize($documentation));
@@ -1819,11 +2131,11 @@ class DocumentationNormalizerTest extends TestCase
         $propertyNameCollectionFactory = $propertyNameCollectionFactoryProphecy->reveal();
         $propertyMetadataFactory = $propertyMetadataFactoryProphecy->reveal();
 
+        $subresourceOperationFactory = new SubresourceOperationFactory($resourceMetadataFactory, $propertyNameCollectionFactory, $propertyMetadataFactory, new UnderscorePathSegmentNameGenerator());
+
         $formatProviderProphecy = $this->prophesize(OperationAwareFormatsProviderInterface::class);
         $formatProviderProphecy->getFormatsFromOperation(Question::class, 'get', OperationType::ITEM)->willReturn(['json' => ['application/json'], 'csv' => ['text/csv']]);
         $formatProviderProphecy->getFormatsFromOperation(Answer::class, 'get', OperationType::SUBRESOURCE)->willReturn(['xml' => ['text/xml']]);
-
-        $subresourceOperationFactory = new SubresourceOperationFactory($resourceMetadataFactory, $propertyNameCollectionFactory, $propertyMetadataFactory, new UnderscorePathSegmentNameGenerator());
 
         $normalizer = new DocumentationNormalizer(
             $resourceMetadataFactory,
@@ -1832,15 +2144,29 @@ class DocumentationNormalizerTest extends TestCase
             $resourceClassResolverProphecy->reveal(),
             $operationMethodResolverProphecy->reveal(),
             $operationPathResolver,
-            null, null, null, false, '', '', '', '', [], [],
+            null,
+            null,
+            null,
+            false,
+            '',
+            '',
+            '',
+            '',
+            [],
+            [],
             $subresourceOperationFactory,
-            true, 'page', false, 'itemsPerPage',
-            $formatProviderProphecy->reveal()
+            true,
+            'page',
+            false,
+            'itemsPerPage',
+            $formatProviderProphecy->reveal(),
+            false,
+            'pagination',
+            ['spec_version' => 3]
         );
 
         $expected = [
-            'swagger' => '2.0',
-            'basePath' => '/',
+            'openapi' => '3.0.2',
             'info' => [
                 'title' => 'Test API',
                 'description' => 'This is a test API.',
@@ -1851,22 +2177,28 @@ class DocumentationNormalizerTest extends TestCase
                     'get' => new \ArrayObject([
                         'tags' => ['Question'],
                         'operationId' => 'getQuestionItem',
-                        'produces' => ['application/json', 'text/csv'],
                         'summary' => 'Retrieves a Question resource.',
                         'parameters' => [
                             [
                                 'name' => 'id',
                                 'in' => 'path',
-                                'type' => 'string',
+                                'schema' => ['type' => 'string'],
                                 'required' => true,
                             ],
                         ],
                         'responses' => [
-                            200 => [
+                            '200' => [
                                 'description' => 'Question resource response',
-                                'schema' => ['$ref' => '#/definitions/Question'],
+                                'content' => [
+                                    'application/json' => [
+                                        'schema' => ['$ref' => '#/components/schemas/Question'],
+                                    ],
+                                    'text/csv' => [
+                                        'schema' => ['$ref' => '#/components/schemas/Question'],
+                                    ],
+                                ],
                             ],
-                            404 => ['description' => 'Resource not found'],
+                            '404' => ['description' => 'Resource not found'],
                         ],
                     ]),
                 ],
@@ -1874,58 +2206,64 @@ class DocumentationNormalizerTest extends TestCase
                     'get' => new \ArrayObject([
                         'tags' => ['Answer', 'Question'],
                         'operationId' => 'api_questions_answer_get_subresource',
-                        'produces' => ['text/xml'],
                         'summary' => 'Retrieves a Answer resource.',
                         'responses' => [
-                            200 => [
+                            '200' => [
                                 'description' => 'Answer resource response',
-                                'schema' => ['$ref' => '#/definitions/Answer'],
+                                'content' => [
+                                    'text/xml' => [
+                                        'schema' => ['$ref' => '#/components/schemas/Answer'],
+                                    ],
+                                ],
                             ],
-                            404 => ['description' => 'Resource not found'],
+                            '404' => ['description' => 'Resource not found'],
                         ],
                         'parameters' => [
                             [
                                 'name' => 'id',
                                 'in' => 'path',
-                                'type' => 'string',
+                                'schema' => ['type' => 'string'],
                                 'required' => true,
                             ],
                         ],
                     ]),
                 ]),
             ]),
-            'definitions' => new \ArrayObject([
-                'Question' => new \ArrayObject([
-                    'type' => 'object',
-                    'description' => 'This is a question.',
-                    'externalDocs' => ['url' => 'http://schema.example.com/Question'],
-                    'properties' => [
-                        'answer' => new \ArrayObject([
-                            'type' => 'array',
-                            'description' => 'This is a name.',
-                            'items' => ['$ref' => '#/definitions/Answer'],
-                        ]),
-                    ],
+            'components' => [
+                'schemas' => new \ArrayObject([
+                    'Question' => new \ArrayObject([
+                        'type' => 'object',
+                        'description' => 'This is a question.',
+                        'externalDocs' => ['url' => 'http://schema.example.com/Question'],
+                        'properties' => [
+                            'answer' => new \ArrayObject([
+                                'type' => 'array',
+                                'description' => 'This is a name.',
+                                'items' => ['$ref' => '#/components/schemas/Answer'],
+                            ]),
+                        ],
+                    ]),
+                    'Answer' => new \ArrayObject([
+                        'type' => 'object',
+                        'description' => 'This is an answer.',
+                        'externalDocs' => ['url' => 'http://schema.example.com/Answer'],
+                        'properties' => [
+                            'content' => new \ArrayObject([
+                                'type' => 'array',
+                                'description' => 'This is a name.',
+                                'items' => ['$ref' => '#/components/schemas/Answer'],
+                            ]),
+                        ],
+                    ]),
                 ]),
-                'Answer' => new \ArrayObject([
-                    'type' => 'object',
-                    'description' => 'This is an answer.',
-                    'externalDocs' => ['url' => 'http://schema.example.com/Answer'],
-                    'properties' => [
-                        'content' => new \ArrayObject([
-                            'type' => 'array',
-                            'description' => 'This is a name.',
-                            'items' => ['$ref' => '#/definitions/Answer'],
-                        ]),
-                    ],
-                ]),
-            ]),
+            ],
+            'servers' => [['url' => '/']],
         ];
 
         $this->assertEquals($expected, $normalizer->normalize($documentation));
     }
 
-    public function testNormalizeWithPropertySwaggerContext()
+    public function testNormalizeWithPropertyOpenApiContext()
     {
         $documentation = new Documentation(new ResourceNameCollection([Dummy::class]), 'Test API', 'This is a test API.', '1.2.3', ['jsonld' => ['application/ld+json']]);
 
@@ -1938,7 +2276,7 @@ class DocumentationNormalizerTest extends TestCase
 
         $propertyMetadataFactoryProphecy = $this->prophesize(PropertyMetadataFactoryInterface::class);
         $propertyMetadataFactoryProphecy->create(Dummy::class, 'id')->shouldBeCalled()->willReturn(new PropertyMetadata(new Type(Type::BUILTIN_TYPE_INT), 'This is an id.', true, false));
-        $propertyMetadataFactoryProphecy->create(Dummy::class, 'name')->shouldBeCalled()->willReturn(new PropertyMetadata(new Type(Type::BUILTIN_TYPE_STRING), 'This is a name.', true, true, true, true, false, false, null, null, ['swagger_context' => ['type' => 'string', 'enum' => ['one', 'two'], 'example' => 'one']]));
+        $propertyMetadataFactoryProphecy->create(Dummy::class, 'name')->shouldBeCalled()->willReturn(new PropertyMetadata(new Type(Type::BUILTIN_TYPE_STRING), 'This is a name.', true, true, true, true, false, false, null, null, ['openapi_context' => ['type' => 'string', 'enum' => ['one', 'two'], 'example' => 'one']]));
         $resourceClassResolverProphecy = $this->prophesize(ResourceClassResolverInterface::class);
         $resourceClassResolverProphecy->isResourceClass(Dummy::class)->willReturn(true);
 
@@ -1953,12 +2291,31 @@ class DocumentationNormalizerTest extends TestCase
             $propertyMetadataFactoryProphecy->reveal(),
             $resourceClassResolverProphecy->reveal(),
             $operationMethodResolverProphecy->reveal(),
-            $operationPathResolver
+            $operationPathResolver,
+            null,
+            null,
+            null,
+            false,
+            '',
+            '',
+            '',
+            '',
+            [],
+            [],
+            null,
+            true,
+            'page',
+            false,
+            'itemsPerPage',
+            null,
+            false,
+            'pagination',
+            ['spec_version' => 3]
         );
 
         $expected = [
-            'swagger' => '2.0',
-            'basePath' => '/app_dev.php/',
+            'openapi' => '3.0.2',
+            'servers' => [['url' => '/app_dev.php/']],
             'info' => [
                 'title' => 'Test API',
                 'description' => 'This is a test API.',
@@ -1969,46 +2326,51 @@ class DocumentationNormalizerTest extends TestCase
                     'get' => new \ArrayObject([
                         'tags' => ['Dummy'],
                         'operationId' => 'getDummyItem',
-                        'produces' => ['application/ld+json'],
                         'summary' => 'Retrieves a Dummy resource.',
                         'parameters' => [
                             [
                                 'name' => 'id',
                                 'in' => 'path',
-                                'type' => 'string',
+                                'schema' => ['type' => 'string'],
                                 'required' => true,
                             ],
                         ],
                         'responses' => [
-                            200 => [
+                            '200' => [
                                 'description' => 'Dummy resource response',
-                                'schema' => ['$ref' => '#/definitions/Dummy'],
+                                'content' => [
+                                    'application/ld+json' => [
+                                        'schema' => ['$ref' => '#/components/schemas/Dummy'],
+                                    ],
+                                ],
                             ],
-                            404 => ['description' => 'Resource not found'],
+                            '404' => ['description' => 'Resource not found'],
                         ],
                     ]),
                 ],
             ]),
-            'definitions' => new \ArrayObject([
-                'Dummy' => new \ArrayObject([
-                    'type' => 'object',
-                    'description' => 'This is a dummy.',
-                    'externalDocs' => ['url' => 'http://schema.example.com/Dummy'],
-                    'properties' => [
-                        'id' => new \ArrayObject([
-                            'type' => 'integer',
-                            'description' => 'This is an id.',
-                            'readOnly' => true,
-                        ]),
-                        'name' => new \ArrayObject([
-                            'type' => 'string',
-                            'description' => 'This is a name.',
-                            'enum' => ['one', 'two'],
-                            'example' => 'one',
-                        ]),
-                    ],
+            'components' => [
+                'schemas' => new \ArrayObject([
+                    'Dummy' => new \ArrayObject([
+                        'type' => 'object',
+                        'description' => 'This is a dummy.',
+                        'externalDocs' => ['url' => 'http://schema.example.com/Dummy'],
+                        'properties' => [
+                            'id' => new \ArrayObject([
+                                'type' => 'integer',
+                                'description' => 'This is an id.',
+                                'readOnly' => true,
+                            ]),
+                            'name' => new \ArrayObject([
+                                'type' => 'string',
+                                'description' => 'This is a name.',
+                                'enum' => ['one', 'two'],
+                                'example' => 'one',
+                            ]),
+                        ],
+                    ]),
                 ]),
-            ]),
+            ],
         ];
 
         $this->assertEquals($expected, $normalizer->normalize($documentation, DocumentationNormalizer::FORMAT, ['base_url' => '/app_dev.php/']));
@@ -2027,7 +2389,7 @@ class DocumentationNormalizerTest extends TestCase
 
         $propertyMetadataFactoryProphecy = $this->prophesize(PropertyMetadataFactoryInterface::class);
         $propertyMetadataFactoryProphecy->create(Dummy::class, 'id')->shouldBeCalled()->willReturn(new PropertyMetadata(new Type(Type::BUILTIN_TYPE_INT), 'This is an id.', true, false));
-        $propertyMetadataFactoryProphecy->create(Dummy::class, 'name')->shouldBeCalled()->willReturn(new PropertyMetadata(new Type(Type::BUILTIN_TYPE_STRING), 'This is a name.', true, true, true, true, false, false, null, null, ['swagger_context' => ['type' => 'string', 'enum' => ['one', 'two'], 'example' => 'one']]));
+        $propertyMetadataFactoryProphecy->create(Dummy::class, 'name')->shouldBeCalled()->willReturn(new PropertyMetadata(new Type(Type::BUILTIN_TYPE_STRING), 'This is a name.', true, true, true, true, false, false, null, null, ['openapi_context' => ['type' => 'string', 'enum' => ['one', 'two'], 'example' => 'one']]));
         $resourceClassResolverProphecy = $this->prophesize(ResourceClassResolverInterface::class);
         $resourceClassResolverProphecy->isResourceClass(Dummy::class)->willReturn(true);
 
@@ -2042,12 +2404,31 @@ class DocumentationNormalizerTest extends TestCase
             $propertyMetadataFactoryProphecy->reveal(),
             $resourceClassResolverProphecy->reveal(),
             $operationMethodResolverProphecy->reveal(),
-            $operationPathResolver
+            $operationPathResolver,
+            null,
+            null,
+            null,
+            false,
+            '',
+            '',
+            '',
+            '',
+            [],
+            [],
+            null,
+            true,
+            'page',
+            false,
+            'itemsPerPage',
+            null,
+            false,
+            'pagination',
+            ['spec_version' => 3]
         );
 
         $expected = [
-            'swagger' => '2.0',
-            'basePath' => '/app_dev.php/',
+            'openapi' => '3.0.2',
+            'servers' => [['url' => '/app_dev.php/']],
             'info' => [
                 'title' => 'Test API',
                 'description' => 'This is a test API.',
@@ -2058,56 +2439,61 @@ class DocumentationNormalizerTest extends TestCase
                     'get' => new \ArrayObject([
                         'tags' => ['Dummy'],
                         'operationId' => 'getDummyCollection',
-                        'produces' => ['application/ld+json'],
                         'summary' => 'Retrieves the collection of Dummy resources.',
                         'parameters' => [
                             [
                                 'name' => 'page',
                                 'in' => 'query',
                                 'required' => false,
-                                'type' => 'integer',
+                                'schema' => ['type' => 'integer'],
                                 'description' => 'The collection page number',
                             ],
                             [
                                 'name' => 'pagination',
                                 'in' => 'query',
                                 'required' => false,
-                                'type' => 'boolean',
+                                'schema' => ['type' => 'boolean'],
                                 'description' => 'Enable or disable pagination',
                             ],
                         ],
                         'responses' => [
-                            200 => [
+                            '200' => [
                                 'description' => 'Dummy collection response',
-                                'schema' => [
-                                    'type' => 'array',
-                                    'items' => ['$ref' => '#/definitions/Dummy'],
+                                'content' => [
+                                    'application/ld+json' => [
+                                        'schema' => [
+                                            'type' => 'array',
+                                            'items' => ['$ref' => '#/components/schemas/Dummy'],
+                                        ],
+                                    ],
                                 ],
                             ],
                         ],
                     ]),
                 ],
             ]),
-            'definitions' => new \ArrayObject([
-                'Dummy' => new \ArrayObject([
-                    'type' => 'object',
-                    'description' => 'This is a dummy.',
-                    'externalDocs' => ['url' => 'http://schema.example.com/Dummy'],
-                    'properties' => [
-                        'id' => new \ArrayObject([
-                            'type' => 'integer',
-                            'description' => 'This is an id.',
-                            'readOnly' => true,
-                        ]),
-                        'name' => new \ArrayObject([
-                            'type' => 'string',
-                            'description' => 'This is a name.',
-                            'enum' => ['one', 'two'],
-                            'example' => 'one',
-                        ]),
-                    ],
+            'components' => [
+                'schemas' => new \ArrayObject([
+                    'Dummy' => new \ArrayObject([
+                        'type' => 'object',
+                        'description' => 'This is a dummy.',
+                        'externalDocs' => ['url' => 'http://schema.example.com/Dummy'],
+                        'properties' => [
+                            'id' => new \ArrayObject([
+                                'type' => 'integer',
+                                'description' => 'This is an id.',
+                                'readOnly' => true,
+                            ]),
+                            'name' => new \ArrayObject([
+                                'type' => 'string',
+                                'description' => 'This is a name.',
+                                'enum' => ['one', 'two'],
+                                'example' => 'one',
+                            ]),
+                        ],
+                    ]),
                 ]),
-            ]),
+            ],
         ];
 
         $this->assertEquals($expected, $normalizer->normalize($documentation, DocumentationNormalizer::FORMAT, ['base_url' => '/app_dev.php/']));
@@ -2151,14 +2537,30 @@ class DocumentationNormalizerTest extends TestCase
             $resourceClassResolverProphecy->reveal(),
             $operationMethodResolverProphecy->reveal(),
             $operationPathResolver,
-            null, null, null, false, '', '', '', '', [], [],
-            null, false, 'page', false, 'itemsPerPage',
-            $formatProviderProphecy->reveal()
+            null,
+            null,
+            null, false,
+            '',
+            '',
+            '',
+            '',
+            [],
+            [],
+
+            null,
+            false,
+            'page',
+            false,
+            'itemsPerPage',
+            $formatProviderProphecy->reveal(),
+            false,
+            'pagination',
+            ['spec_version' => 3]
         );
 
         $expected = [
-            'swagger' => '2.0',
-            'basePath' => '/',
+            'openapi' => '3.0.2',
+            'servers' => [['url' => '/']],
             'info' => [
                 'title' => 'Test API',
                 'description' => 'This is a test API.',
@@ -2169,15 +2571,24 @@ class DocumentationNormalizerTest extends TestCase
                     'get' => new \ArrayObject([
                         'tags' => ['Dummy'],
                         'operationId' => 'getDummyCollection',
-                        'produces' => ['application/xml', 'text/xml'],
                         'summary' => 'Retrieves the collection of Dummy resources.',
                         'parameters' => [],
                         'responses' => [
                             200 => [
                                 'description' => 'Dummy collection response',
-                                'schema' => [
-                                    'type' => 'array',
-                                    'items' => ['$ref' => '#/definitions/Dummy'],
+                                'content' => [
+                                    'application/xml' => [
+                                        'schema' => [
+                                            'type' => 'array',
+                                            'items' => ['$ref' => '#/components/schemas/Dummy'],
+                                        ],
+                                    ],
+                                    'text/xml' => [
+                                        'schema' => [
+                                            'type' => 'array',
+                                            'items' => ['$ref' => '#/components/schemas/Dummy'],
+                                        ],
+                                    ],
                                 ],
                             ],
                         ],
@@ -2185,21 +2596,29 @@ class DocumentationNormalizerTest extends TestCase
                     'post' => new \ArrayObject([
                         'tags' => ['Dummy'],
                         'operationId' => 'postDummyCollection',
-                        'consumes' => ['text/xml', 'text/csv'],
-                        'produces' => ['text/xml', 'text/csv'],
                         'summary' => 'Creates a Dummy resource.',
-                        'parameters' => [
-                            [
-                                'name' => 'dummy',
-                                'in' => 'body',
-                                'description' => 'The new Dummy resource',
-                                'schema' => ['$ref' => '#/definitions/Dummy'],
+                        'requestBody' => [
+                            'content' => [
+                                'text/xml' => [
+                                    'schema' => ['$ref' => '#/components/schemas/Dummy'],
+                                ],
+                                'text/csv' => [
+                                    'schema' => ['$ref' => '#/components/schemas/Dummy'],
+                                ],
                             ],
+                            'description' => 'The new Dummy resource',
                         ],
                         'responses' => [
                             201 => [
                                 'description' => 'Dummy resource created',
-                                'schema' => ['$ref' => '#/definitions/Dummy'],
+                                'content' => [
+                                    'text/xml' => [
+                                        'schema' => ['$ref' => '#/components/schemas/Dummy'],
+                                    ],
+                                    'text/csv' => [
+                                        'schema' => ['$ref' => '#/components/schemas/Dummy'],
+                                    ],
+                                ],
                             ],
                             400 => ['description' => 'Invalid input'],
                             404 => ['description' => 'Resource not found'],
@@ -2210,20 +2629,23 @@ class DocumentationNormalizerTest extends TestCase
                     'get' => new \ArrayObject([
                         'tags' => ['Dummy'],
                         'operationId' => 'getDummyItem',
-                        'produces' => ['application/vnd.api+json'],
                         'summary' => 'Retrieves a Dummy resource.',
                         'parameters' => [
                             [
                                 'name' => 'id',
                                 'in' => 'path',
-                                'type' => 'string',
+                                'schema' => ['type' => 'string'],
                                 'required' => true,
                             ],
                         ],
                         'responses' => [
                             200 => [
                                 'description' => 'Dummy resource response',
-                                'schema' => ['$ref' => '#/definitions/Dummy'],
+                                'content' => [
+                                    'application/vnd.api+json' => [
+                                        'schema' => ['$ref' => '#/components/schemas/Dummy'],
+                                    ],
+                                ],
                             ],
                             404 => ['description' => 'Resource not found'],
                         ],
@@ -2231,27 +2653,37 @@ class DocumentationNormalizerTest extends TestCase
                     'put' => new \ArrayObject([
                         'tags' => ['Dummy'],
                         'operationId' => 'putDummyItem',
-                        'consumes' => ['application/json', 'text/csv'],
-                        'produces' => ['application/json', 'text/csv'],
                         'summary' => 'Replaces the Dummy resource.',
                         'parameters' => [
                             [
                                 'name' => 'id',
                                 'in' => 'path',
-                                'type' => 'string',
+                                'schema' => ['type' => 'string'],
                                 'required' => true,
                             ],
-                            [
-                                'name' => 'dummy',
-                                'in' => 'body',
-                                'description' => 'The updated Dummy resource',
-                                'schema' => ['$ref' => '#/definitions/Dummy'],
+                        ],
+                        'requestBody' => [
+                            'content' => [
+                                'application/json' => [
+                                    'schema' => ['$ref' => '#/components/schemas/Dummy'],
+                                ],
+                                'text/csv' => [
+                                    'schema' => ['$ref' => '#/components/schemas/Dummy'],
+                                ],
                             ],
+                            'description' => 'The updated Dummy resource',
                         ],
                         'responses' => [
                             200 => [
                                 'description' => 'Dummy resource updated',
-                                'schema' => ['$ref' => '#/definitions/Dummy'],
+                                'content' => [
+                                    'application/json' => [
+                                        'schema' => ['$ref' => '#/components/schemas/Dummy'],
+                                    ],
+                                    'text/csv' => [
+                                        'schema' => ['$ref' => '#/components/schemas/Dummy'],
+                                    ],
+                                ],
                             ],
                             400 => ['description' => 'Invalid input'],
                             404 => ['description' => 'Resource not found'],
@@ -2259,24 +2691,26 @@ class DocumentationNormalizerTest extends TestCase
                     ]),
                 ],
             ]),
-            'definitions' => new \ArrayObject([
-                'Dummy' => new \ArrayObject([
-                    'type' => 'object',
-                    'description' => 'This is a dummy.',
-                    'externalDocs' => ['url' => 'http://schema.example.com/Dummy'],
-                    'properties' => [
-                        'id' => new \ArrayObject([
-                            'type' => 'integer',
-                            'description' => 'This is an id.',
-                            'readOnly' => true,
-                        ]),
-                        'name' => new \ArrayObject([
-                            'type' => 'string',
-                            'description' => 'This is a name.',
-                        ]),
-                    ],
+            'components' => [
+                'schemas' => new \ArrayObject([
+                    'Dummy' => new \ArrayObject([
+                        'type' => 'object',
+                        'description' => 'This is a dummy.',
+                        'externalDocs' => ['url' => 'http://schema.example.com/Dummy'],
+                        'properties' => [
+                            'id' => new \ArrayObject([
+                                'type' => 'integer',
+                                'description' => 'This is an id.',
+                                'readOnly' => true,
+                            ]),
+                            'name' => new \ArrayObject([
+                                'type' => 'string',
+                                'description' => 'This is a name.',
+                            ]),
+                        ],
+                    ]),
                 ]),
-            ]),
+            ],
         ];
 
         $this->assertEquals($expected, $normalizer->normalize($documentation, DocumentationNormalizer::FORMAT, ['base_url' => '/']));
